fix(kernels): cast_f32_to_bf16/f16 use _rn (round-to-nearest)
The activation-cast kernels used __float2bfloat16 / __float2half (truncating
toward zero) while all typed kernels' from_f_bf16 / from_f_f16 downcasts in
_typed_prelude.cuh use _rn (round-to-nearest). Mismatched rounding adds a
systematic negative bias to any f32→bf16 activation crossing this kernel,
which compounds through successive GEMMs.

The cast kernel is on the hot path when an f32 activation has to be fed into
a bf16-weight GEMM (gpu_gemm_typed_forward wrapper in blas.rs). Found while
investigating 130m bf16 greedy-decoding regression against HF transformers
(still diverges after this fix — the 130m issue is separate; see
internal/bf16_130m_bug.md).

+++ kernels/elementwise.cu
@@ -19,7 +19,12 @@ extern "C" __global__ void cast_f32_to_bf16(
 ) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i >= n) return;
-    dst[i] = __float2bfloat16(src[i]);
+    // Round-to-nearest, matching the intermediate-value downcasts in all
+    // typed kernels (from_f_bf16 in _typed_prelude.cuh). The default
+    // `__float2bfloat16` is round-toward-zero, which adds a systematic
+    // negative bias to every weight and compounds across GEMMs — visible
+    // as degenerate greedy decoding on small models (e.g. mamba-130m).
+    dst[i] = __float2bfloat16_rn(src[i]);
 }
 
 extern "C" __global__ void cast_f32_to_f16(
@@ -27,7 +32,9 @@ extern "C" __global__ void cast_f32_to_f16(
 ) {
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i >= n) return;
-    dst[i] = __float2half(src[i]);
+    // Same rationale as cast_f32_to_bf16: match the _rn rounding mode used
+    // by from_f_f16 in _typed_prelude.cuh.
+    dst[i] = __float2half_rn(src[i]);
 }
 
 extern "C" __global__ void bias_broadcast(
